extract makeRoadLabels
